gavneco

+++ models/user.go
@@ -3,5 +3,5 @@ package models
 type User struct {
 	Id   int64
 	Name string
-	Hp   int
+	Hp   int64
 }
